# table: Always check key ordering when inserting to an SST