To prevent an out-of-order SSTable insertion from ever happening, this
change updates BlockBasedTableBuilder::Add to always assert on key
ordering being maintained at the insertion point, rather than only
running with debug builds. When an out-of-order key insertion is
detected, the status is set to Corrupted to prevent the BlockBasedTable
from being written.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/rocksdb/46)
<!-- Reviewable:end -->