@itsbilal 2019-09-09
Note that there's a significant performance regression with running the comparator on every SST write with this change:

```
name                     old time/op   new time/op   delta
RocksDBSstFileWriter-16    761ns ± 2%    835ns ± 1%  +9.75%  (p=0.008 n=5+5)

name                     old speed     new speed     delta
RocksDBSstFileWriter-16  144MB/s ± 2%  132MB/s ± 1%  -8.87%  (p=0.008 n=5+5)
```

Open to ideas on reducing this, if that's seen as a priority. Are SST compactions / writes ever on a write workload hot path?