@itsbilal 2019-09-09
@petermattis Weird. I ran them on an AWS `c5d.4xlarge` created using roachprod, but good catch - I'll re-run everything. There's a chance something else was happening on the node at that time.

As for the status thing, all the Write* methods seem to gate actual writes on `ok()` returning true.