@petermattis 2019-09-09
> As for the status thing, all the Write* methods seem to gate actual writes on ok() returning true.

It is more than just gating the Write* methods. It is also making sure that the error paths from a failed compaction works. @ajkr knows this code much better than I do, so if he signs off on this approach I'm good with it too.