@itsbilal 2019-09-10
@petermattis I ran the benchmark again, on another node (same AWS type), and got this difference:

```
name                     old time/op   new time/op   delta
RocksDBSstFileWriter-16    722ns ± 2%    779ns ± 1%  +7.90%  (p=0.008 n=5+5)

name                     old speed     new speed     delta
RocksDBSstFileWriter-16  152MB/s ± 2%  141MB/s ± 1%  -7.29%  (p=0.008 n=5+5)
```

I also got a similar difference on my laptop. Maybe cmake didn't actually rebuild rocksdb for your run?