@petermattis 2019-09-10
> I also got a similar difference on my laptop. Maybe cmake didn't actually rebuild rocksdb for your run?

I was careful to bump the value in `c-deps/rocksdb-rebuild` and saw RocksDB rebuilding. Oh well. I think it is worthwhile to turn this on (i.e. merge this PR), but I think we should run one of the write-heavy roachtests to see if there is an impact. The single-node `kv0` roachtests are probably good.