@itsbilal 2019-09-10
Sounds good. I did notice that runtimes on my laptop are much more variable, which seems to overshadow the time delta. Will run those roachtests and see what happens, thanks!